Adds mapping nodes

+++ kimchi_state/include/kimchi_state/map_info.h
@@ -1,14 +1,15 @@
 #pragma once
-#include <rclcpp/rclcpp.hpp>
-#include <vector>
 #include <cstdint>
 #include <geometry_msgs/msg/pose2_d.hpp>
-
+#include <rclcpp/rclcpp.hpp>
+#include <vector>
 
 struct MapInfo {
-    MapInfo(float resolution, const geometry_msgs::msg::Pose2D& origin_, const std::vector<long int>& image ): resolution_(resolution), origin_(origin_), image_(image) {}
+  MapInfo(float resolution, const geometry_msgs::msg::Pose2D& origin_,
+          const std::vector<uint8_t>& image)
+      : resolution_(resolution), origin_(origin_), image_(image) {}
 
-    float resolution_;
-    geometry_msgs::msg::Pose2D origin_;
-    std::vector<long int> image_; // Assuming image is a byte array
+  float resolution_;
+  geometry_msgs::msg::Pose2D origin_;
+  std::vector<uint8_t> image_;  // Assuming image is a byte array
 };
